pyproject.toml
[project]
name = "dbt-starter-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "dbt-core>=1.11.3",
    "dbt-snowflake>=1.11.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "yamllint>=1.35.0",
    "sqlfluff>=3.0.0",
    "sqlfluff-templater-dbt>=4.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
